Wrap LLVM Intrinsics
The LLVM Intrinsics API is large, complex, platform-dependent,
and in a constant state of flux.  LLVM has chosen to manage this
complexity with a set of TableGen definition files.  Therefore,
we must parse and interpret these files to provide a robust API.

This commit introduces a new tool, 'intrinsics-gen', to be invoked
from the top-level of the repository. The tool will read in any number
of platform-dependent TableGen files and interpret them as a Swift file
of enums with selectors.  Each of these selectors either represents an
intrinsic or a family of intrinsics that the framework is capable of
resolving to one of LLVM's "proper intrinsic" selectors.

The idea is that we do not ever ship this generated Swift file.
Instead, users that wish to use this framework with the intrinsics will
invoke 'intrinsics-gen' with the appropriate td file and SwiftPM will
handle the rest.  The API that will be committed is just the
infrastructure necessary to support this.  See the test suite for
a concrete example of how to use this.