Introduce IRInstruction
In order to unify the many structures we have that each represent instruction values, introduce the IRInstruction protocol.  LLVMValueRef now unconditionally conforms to IRInstruction to facilitate dynamic casts.  Note that this is strictly not correct.  In particular, constant folding means the result of a lot of IRBuilder calls wind up as just plain values.  Therefore, the return types for IRBuilder functions has only been refined in as many places where we can guarantee constant-folding will always produce an instruction or instruction sequence.

Ultimately, this protocol greases the wheels for a more robust AttachedMetadata API for which we need new shims.